requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121  # CUDA 12.1 support for torch and torchvision.
# --extra-index-url https://download.pytorch.org/whl/cu118  # CUDA 11.8 support for torch and torchvision.

h5py==3.11.0
jupyter==1.0.0  # Jupyter system, including the notebook and the IPython kernel, all in one go.
lightning==2.3.3
matplotlib==3.9.1
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
tabulate==0.9.0
torch==2.3.1
tqdm==4.66.4
torchmetrics==1.4.0
torchvision==0.18.1
